requirements.txt
streamlit
PyPDF2
langchain
langchain-google-genai
faiss-cpu
google-generativeai
python-dotenv
